+++ neoforge/src/main/kotlin/notlown/cobblebase/neoforge/client/gui/SkillAssignmentScreen.kt
@@ -1,5 +1,6 @@
 package notlown.cobblebase.neoforge.client.gui
 
+import notlown.cobblebase.core.AssignmentCache
 import notlown.cobblebase.core.BaseManager
 import notlown.cobblebase.core.SkillRegistry
 import notlown.cobblebase.core.SpeciesSkillRegistry
@@ -68,7 +69,7 @@ class SkillAssignmentScreen(
         pokemonList.forEachIndexed { index, pokemonData ->
             val pokemonId = pokemonData.pokemonId
             val speciesName = pokemonData.species.path
-            val currentAssignment = BaseManager.getAssignment(pokemonId)
+            val currentAssignment = AssignmentCache.getAssignment(pokemonId)
             val speciesSkills = SpeciesSkillRegistry.getSkills(speciesName)
             val availableSkills = speciesSkills?.skills ?: emptyList()